+++ packages/react-container/src/components/Sentry.tsx
@@ -1,22 +1,25 @@
 import React from 'react';
 import { Sentry as SentryComponent, SentryProps } from '@cognite/react-sentry';
-import { History } from 'history';
 
 import { ConditionalWrapperWithProps } from './ConditionalWrapper';
 
-export const Sentry: React.FC<SentryProps> = ({ children, history }) => (
-  <SentryComponent history={history}>
+export const Sentry: React.FC<SentryProps> = ({
+  children,
+  ...sentrySettings
+}) => (
+  <SentryComponent {...sentrySettings}>
     <>{children}</>
   </SentryComponent>
 );
 
-export const ConditionalSentry: React.FC<{
-  children: React.ReactElement;
-  disabled?: boolean;
-  history: History;
-}> = ({ disabled, children }) => {
+export const ConditionalSentry: React.FC<
+  {
+    children: React.ReactElement;
+    disabled?: boolean;
+  } & SentryProps
+> = ({ disabled, children, ...rest }) => {
   return (
-    <ConditionalWrapperWithProps condition={!disabled} wrap={Sentry}>
+    <ConditionalWrapperWithProps condition={!disabled} wrap={Sentry} {...rest}>
       {children}
     </ConditionalWrapperWithProps>
   );